+++ LosYuyitos/Models/FAMILIAPRODUCTO.cs
@@ -16,7 +16,7 @@ namespace LosYuyitos.Models
         }
 
         [Key]
-        public int TIPOPRODUCTOID { get; set; }
+        public decimal TIPOPRODUCTOID { get; set; }
 
         [StringLength(50)]
         public string DESCRIPCION { get; set; }
